Moved the UserCreateResource class into the util package
