+++ src/main/java/com/aat/utils/ResourceUtil.java
@@ -1,4 +1,4 @@
-package com.aat.restlet;
+package com.aat.utils;
 
 import org.restlet.data.Form;
 import org.restlet.resource.ResourceException;
